requirements.txt
albumentations==1.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cvlib==0.2.7
idna==3.4
imageio==2.28.1
imutils==0.5.4
joblib==1.2.0
networkx==2.6.3
numpy==1.21.6
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
progressbar==2.5
pyserial==3.5
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.30.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
typing-extensions==4.5.0
urllib3==2.0.2
